// lib/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A (potentially binary) string encoded using base64. */
  Base64String: any;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** A string containing HTML code. */
  HTML: any;
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: any;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type DependencyGraphDependency = {
  readonly __typename?: 'DependencyGraphDependency';
  /** Does the dependency itself have dependencies? */
  readonly hasDependencies: Scalars['Boolean'];
  /** The original name of the package, as it appears in the manifest. */
  readonly packageLabel: Scalars['String'];
  /** The dependency package manager */
  readonly packageManager?: Maybe<Scalars['String']>;
  /** The name of the package in the canonical form used by the package manager. This may differ from the original textual form (see packageLabel), for example in a package manager that uses case-insensitive comparisons. */
  readonly packageName: Scalars['String'];
  /** The repository containing the package */
  readonly repository?: Maybe<Repository>;
  /** The dependency version requirements */
  readonly requirements: Scalars['String'];
};

/** The connection type for DependencyGraphDependency. */
export type DependencyGraphDependencyConnection = {
  readonly __typename?: 'DependencyGraphDependencyConnection';
  /** A list of edges. */
  readonly edges?: Maybe<ReadonlyArray<Maybe<DependencyGraphDependencyEdge>>>;
  /** A list of nodes. */
  readonly nodes?: Maybe<ReadonlyArray<Maybe<DependencyGraphDependency>>>;
  /** Information to aid in pagination. */
  readonly pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  readonly totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type DependencyGraphDependencyEdge = {
  readonly __typename?: 'DependencyGraphDependencyEdge';
  /** A cursor for use in pagination. */
  readonly cursor: Scalars['String'];
  /** The item at the end of the edge. */
  readonly node?: Maybe<DependencyGraphDependency>;
};

/** Dependency manifest for a repository */
export type DependencyGraphManifest = Node & {
  readonly __typename?: 'DependencyGraphManifest';
  /** Path to view the manifest file blob */
  readonly blobPath: Scalars['String'];
  /** A list of manifest dependencies */
  readonly dependencies?: Maybe<DependencyGraphDependencyConnection>;
  /** The number of dependencies listed in the manifest */
  readonly dependenciesCount?: Maybe<Scalars['Int']>;
  /** Is the manifest too big to parse? */
  readonly exceedsMaxSize: Scalars['Boolean'];
  /** Fully qualified manifest filename */
  readonly filename: Scalars['String'];
  readonly id: Scalars['ID'];
  /** Were we able to parse the manifest? */
  readonly parseable: Scalars['Boolean'];
  /** The repository containing the manifest */
  readonly repository: Repository;
};


/** Dependency manifest for a repository */
export type DependencyGraphManifestDependenciesArgs = {
  after?: InputMaybe<Scalars['String']>;
  before?: InputMaybe<Scalars['String']>;
  first?: InputMaybe<Scalars['Int']>;
  last?: InputMaybe<Scalars['Int']>;
};

/** The connection type for DependencyGraphManifest. */
export type DependencyGraphManifestConnection = {
  readonly __typename?: 'DependencyGraphManifestConnection';
  /** A list of edges. */
  readonly edges?: Maybe<ReadonlyArray<Maybe<DependencyGraphManifestEdge>>>;
  /** A list of nodes. */
  readonly nodes?: Maybe<ReadonlyArray<Maybe<DependencyGraphManifest>>>;
  /** Information to aid in pagination. */
  readonly pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  readonly totalCount: Scalars['Int'];
};
// ...
export type Node = {
  /** ID of the object. */
  readonly id: Scalars['ID'];
};
// ...
export type PageInfo = {
  readonly __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  readonly endCursor?: Maybe<Scalars['String']>;
  /** When paginating forwards, are there more items? */
  readonly hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  readonly hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  readonly startCursor?: Maybe<Scalars['String']>;
};
// ...
export type Query = {
  readonly __typename?: 'Query';
  /** Look up a code of conduct by its key */
  readonly codeOfConduct?: Maybe<CodeOfConduct>;
  /** Look up a code of conduct by its key */
  readonly codesOfConduct?: Maybe<ReadonlyArray<Maybe<CodeOfConduct>>>;
  /** Look up an enterprise by URL slug. */
  readonly enterprise?: Maybe<Enterprise>;
  /** Look up a pending enterprise administrator invitation by invitee, enterprise and role. */
  readonly enterpriseAdministratorInvitation?: Maybe<EnterpriseAdministratorInvitation>;
  /** Look up a pending enterprise administrator invitation by invitation token. */
  readonly enterpriseAdministratorInvitationByToken?: Maybe<EnterpriseAdministratorInvitation>;
  /** Look up an open source license by its key */
  readonly license?: Maybe<License>;
  /** Return a list of known open source licenses */
  readonly licenses: ReadonlyArray<Maybe<License>>;
  /** Get alphabetically sorted list of Marketplace categories */
  readonly marketplaceCategories: ReadonlyArray<MarketplaceCategory>;
  /** Look up a Marketplace category by its slug. */
  readonly marketplaceCategory?: Maybe<MarketplaceCategory>;
  /** Look up a single Marketplace listing */
  readonly marketplaceListing?: Maybe<MarketplaceListing>;
  /** Look up Marketplace listings */
  readonly marketplaceListings: MarketplaceListingConnection;
  /** Return information about the GitHub instance */
  readonly meta: GitHubMetadata;
  /** Fetches an object given its ID. */
  readonly node?: Maybe<Node>;
  /** Lookup nodes by a list of IDs. */
  readonly nodes: ReadonlyArray<Maybe<Node>>;
  /** Lookup a organization by login. */
  readonly organization?: Maybe<Organization>;
  /** The client's rate limit information. */
  readonly rateLimit?: Maybe<RateLimit>;
  /** Hack to workaround https://github.com/facebook/relay/issues/112 re-exposing the root query object */
  readonly relay: Query;
  /** Lookup a given repository by the owner and repository name. */
  readonly repository?: Maybe<Repository>;
  /** Lookup a repository owner (ie. either a User or an Organization) by login. */
  readonly repositoryOwner?: Maybe<RepositoryOwner>;
  /** Lookup resource by a URL. */
  readonly resource?: Maybe<UniformResourceLocatable>;
  /** Perform a search across resources. */
  readonly search: SearchResultItemConnection;
  /** GitHub Security Advisories */
  readonly securityAdvisories: SecurityAdvisoryConnection;
  /** Fetch a Security Advisory by its GHSA ID */
  readonly securityAdvisory?: Maybe<SecurityAdvisory>;
  /** Software Vulnerabilities documented by GitHub Security Advisories */
  readonly securityVulnerabilities: SecurityVulnerabilityConnection;
  /** Users and organizations who can be sponsored via GitHub Sponsors. */
  readonly sponsorables: SponsorableItemConnection;
  /** Look up a topic by name. */
  readonly topic?: Maybe<Topic>;
  /** Lookup a user by login. */
  readonly user?: Maybe<User>;
  /** The currently authenticated user. */
  readonly viewer: User;
};
// ...
export type RepositoryConnection = {
  readonly __typename?: 'RepositoryConnection';
  /** A list of edges. */
  readonly edges?: Maybe<ReadonlyArray<Maybe<RepositoryEdge>>>;
  /** A list of nodes. */
  readonly nodes?: Maybe<ReadonlyArray<Maybe<Repository>>>;
  /** Information to aid in pagination. */
  readonly pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  readonly totalCount: Scalars['Int'];
  /** The total size in kilobytes of all repositories in the connection. */
  readonly totalDiskUsage: Scalars['Int'];
};
// ...
export type RepositoryEdge = {
  readonly __typename?: 'RepositoryEdge';
  /** A cursor for use in pagination. */
  readonly cursor: Scalars['String'];
  /** The item at the end of the edge. */
  readonly node?: Maybe<Repository>;
};
// ...
export type ManifestFragment = { readonly __typename?: 'DependencyGraphManifest', readonly id: string, readonly filename: string, readonly dependenciesCount?: number | null, readonly dependencies?: { readonly __typename?: 'DependencyGraphDependencyConnection', readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphDependency', readonly packageName: string, readonly packageManager?: string | null, readonly requirements: string } | null> | null } | null };

export type DependencyFragment = { readonly __typename?: 'DependencyGraphDependency', readonly packageName: string, readonly packageManager?: string | null, readonly requirements: string };

export type RepositoryFragment = { readonly __typename?: 'Repository', readonly name: string, readonly nameWithOwner: string, readonly dependencyGraphManifests?: { readonly __typename?: 'DependencyGraphManifestConnection', readonly totalCount: number, readonly pageInfo: { readonly __typename?: 'PageInfo', readonly hasNextPage: boolean }, readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphManifest', readonly id: string, readonly filename: string, readonly dependenciesCount?: number | null, readonly dependencies?: { readonly __typename?: 'DependencyGraphDependencyConnection', readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphDependency', readonly packageName: string, readonly packageManager?: string | null, readonly requirements: string } | null> | null } | null } | null> | null } | null };

export type AllReposQueryVariables = Exact<{
  login: Scalars['String'];
  first: Scalars['Int'];
  after?: InputMaybe<Scalars['String']>;
}>;


export type AllReposQuery = { readonly __typename?: 'Query', readonly organization?: { readonly __typename?: 'Organization', readonly name?: string | null, readonly repositories: { readonly __typename?: 'RepositoryConnection', readonly totalCount: number, readonly pageInfo: { readonly __typename?: 'PageInfo', readonly hasNextPage: boolean }, readonly edges?: ReadonlyArray<{ readonly __typename?: 'RepositoryEdge', readonly cursor: string, readonly node?: { readonly __typename?: 'Repository', readonly name: string, readonly nameWithOwner: string, readonly dependencyGraphManifests?: { readonly __typename?: 'DependencyGraphManifestConnection', readonly totalCount: number, readonly pageInfo: { readonly __typename?: 'PageInfo', readonly hasNextPage: boolean }, readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphManifest', readonly id: string, readonly filename: string, readonly dependenciesCount?: number | null, readonly dependencies?: { readonly __typename?: 'DependencyGraphDependencyConnection', readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphDependency', readonly packageName: string, readonly packageManager?: string | null, readonly requirements: string } | null> | null } | null } | null> | null } | null } | null } | null> | null } } | null };

export type PagedManifestsQueryVariables = Exact<{
  login: Scalars['String'];
  repoName: Scalars['String'];
  after?: InputMaybe<Scalars['String']>;
  first: Scalars['Int'];
}>;


export type PagedManifestsQuery = { readonly __typename?: 'Query', readonly organization?: { readonly __typename?: 'Organization', readonly name?: string | null, readonly repository?: { readonly __typename?: 'Repository', readonly name: string, readonly nameWithOwner: string, readonly dependencyGraphManifests?: { readonly __typename?: 'DependencyGraphManifestConnection', readonly totalCount: number, readonly pageInfo: { readonly __typename?: 'PageInfo', readonly hasNextPage: boolean }, readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphManifest', readonly id: string, readonly filename: string, readonly dependenciesCount?: number | null, readonly dependencies?: { readonly __typename?: 'DependencyGraphDependencyConnection', readonly nodes?: ReadonlyArray<{ readonly __typename?: 'DependencyGraphDependency', readonly packageName: string, readonly packageManager?: string | null, readonly requirements: string } | null> | null } | null } | null> | null } | null } | null } | null };